requirements.txt
alembic==1.4.2
aniso8601==6.0.0
astroid==2.2.5
atomicwrites==1.4.0
attrs==20.1.0
awscli==1.18.127
botocore==1.17.50
Click==7.0
colorama==0.4.3
docutils==0.15.2
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==20.0.4
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.2
py==1.9.0
pyasn1==0.4.8
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.1
typed-ast==1.3.5
urllib3==1.25.10
Werkzeug==0.15.2
wrapt==1.11.1
zipp==3.1.0
